create prompts
