+++ src/runners/commander/types.ts
@@ -1,0 +1,6 @@
+export type CliOptions = {
+  template: string;
+  importAlias: string;
+  packageManager: string;
+  resetPreferences: boolean;
+};
